changed return type of getReadByteCount method in Unpacker interface: long to int

+++ src/main/java/org/msgpack/io/Input.java
@@ -42,7 +42,7 @@ public interface Input extends Closeable {
 
     public double getDouble() throws IOException;
 
-    public long getReadByteCount();
+    public int getReadByteCount();
 
     public void resetReadByteCount();
 }
